ejemplo_pytest_sin_bdd/requirements.txt
psycopg2-binary
selenium
webdriver-manager
pytest
pytest-xdist
allure-pytest
filelock
pytest-percent